requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0
opencv-python==4.8.1.78
nltk==3.8.1
scikit-learn==1.3.2
python-multipart==0.0.6
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
plotly==5.18.0
matplotlib==3.8.2 